delphi: premier test réussi tcOpenDocument_Embed_Image (nécessite de mettre un 7z.dll 32bits dans pascal_o_r_mapping\delphi\DUnitX\duxPASCAL_O_R_MAPPING\Win32\Debug)

+++ delphi/pascal_o_r_mapping/2_Units/uProgression.pas
@@ -26,13 +26,13 @@ interface
 
 var
    uProgression_Demarre        : procedure ( Titre: String;
-                                             Min, Max: Integer;
+                                             Min, Max: Single;
                                              _Immediat: Boolean = False;
                                              _Interruptible: Boolean= False) of object = nil;
-   uProgression_DemarreImmediat: procedure ( Titre: String; Min, Max: Integer;
+   uProgression_DemarreImmediat: procedure ( Titre: String; Min, Max: Single;
                                              _Interruptible: Boolean= False) of object = nil;
    uProgression_Termine        : procedure of object = nil;
-   uProgression_AddProgress    : procedure ( Value: Integer) of object= nil;
+   uProgression_AddProgress    : procedure ( Value: Single) of object= nil;
    uProgression_GetInterrompre : function : Boolean of object= nil;
 
 implementation
